third_party/blink/web_tests/virtual/dom-parts-minimal/external/wpt/dom/parts/basic-dom-part-declarative-brace-syntax-innerhtml.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] Basic AttributePart parsing with multiple attributes
  assert_equals: Declarative AttributePart should be present: lengths differ expected 2 but got 0
Harness: the test ran to completion.
